Fix build warnings

+++ include/dsp_base.h
@@ -3,7 +3,7 @@
 // =============================================================================
 #pragma once
 
-#define TWO_PI (2.0f * M_PI)
+#define TWO_PI static_cast<float>((2.0f * M_PI))
 
 namespace dsp
 {
